% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildpathFileAndNameValid.R
\name{buildpathFileAndNameValid}
\alias{buildpathFileAndNameValid}
\title{creates the full path to an existing file in a directory}
\usage{
buildpathFileAndNameValid(pathFile, nameFileImag)
}
\arguments{
\item{pathFile}{: string (path)}

\item{nameFileImag}{: string (file name)}
}
\value{
NULL or string (full path)
}
\description{
creates the full path to an existing file in a directory
}
\examples{
pathFile <- "C:/Images"
nameFileImag <- "Image.bmp"
buildpathFileAndNameValid(pathFile, nameFileImag)

}
