% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImgAllInfoData.R
\name{ImgAllInfoData}
\alias{ImgAllInfoData}
\title{gets the data from a raster images set}
\usage{
ImgAllInfoData(pathFile, pathFileCsv, nameFileCsv = "File_001.csv",
  saveCsvFile = TRUE, withGraphic = TRUE)
}
\arguments{
\item{pathFile}{: string (path)}

\item{pathFileCsv}{: string (path CSV file)}

\item{nameFileCsv}{: string (name CSV file)}

\item{saveCsvFile}{: TRUE or FALSE}

\item{withGraphic}{: TRUE or FALSE}
}
\value{
data.frame file
}
\description{
gets the data from a raster images set
}
\examples{
pathFile <- "C:/Images"
pathFileCsv <- "C:/Csv"
nameFileCsv <- "ImgDataX.csv"
dataFrame <- ImgAllInfoData(pathFile, pathFileCsv, nameFileCsv)
}
