% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgEntropy.R
\name{imgEntropy}
\alias{imgEntropy}
\title{gets the entropy value from image file}
\usage{
imgEntropy(pathFile, nameFileImag01)
}
\arguments{
\item{pathFile}{: string (path)}

\item{nameFileImag01}{: string (file name)}
}
\value{
number
}
\description{
gets the entropy value from image file
}
\examples{
pathFile <- "C:/Images"
nameFileImag01 <- "Img01.bmp"
imgEntropy(pathFile, nameFileImag01)

}
