% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgHistogram.R
\name{imgHistogram}
\alias{imgHistogram}
\title{shows the image histogram}
\usage{
imgHistogram(pathFile, nameFileImag)
}
\arguments{
\item{pathFile}{: string (path)}

\item{nameFileImag}{: string (file name)}
}
\description{
shows the image histogram
}
\examples{

pathFile <- "C:/Images"
nameFileImag <- "Image.bmp"
imgHistogram(pathFile, nameFileImag)

}
