% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgInfoData.R
\name{imgInfoData}
\alias{imgInfoData}
\title{gets the data from a raster image}
\usage{
imgInfoData(pathFile, nameFileImag01, viewHist = TRUE,
  viewDatArray = TRUE)
}
\arguments{
\item{pathFile}{: string (path)}

\item{nameFileImag01}{:  string (name image file)}

\item{viewHist}{: TRUE or FALSE (show histogram)}

\item{viewDatArray}{: TRUE or FALSE (show show data.frame)}
}
\value{
vector of numbers
}
\description{
gets the data from a raster image
}
\examples{
pathFile <- "C:/Images"
nameFileImag01 <- "ImgDataX.csv"
imgInfoData <- function(pathFile, nameFileImag01)
}
