% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgRangeMinMax.R
\name{imgRangeMinMax}
\alias{imgRangeMinMax}
\title{gets the values range from image file}
\usage{
imgRangeMinMax(pathFile, nameFileImag01)
}
\arguments{
\item{pathFile}{: string (path)}

\item{nameFileImag01}{: string (file name)}
}
\value{
vector of real numbers
}
\description{
gets the values range from image file
}
\examples{

pathFile <- "C:/Images"
nameFileImag01 <- "Image.bmp"
imgRangeMinMax(pathFile, nameFileImag01)

}
