% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgStDeviation.R
\name{imgStDeviation}
\alias{imgStDeviation}
\title{gets the standard deviation value from image file}
\usage{
imgStDeviation(pathFile, nameFileImag01)
}
\arguments{
\item{pathFile}{: string (path)}

\item{nameFileImag01}{: string (file name)}
}
\value{
number
}
\description{
gets the standard deviation value from image file
}
\examples{

pathFile <- "C:/Images"
nameFileImag01 <- "Image.bmp"
imgStDeviation(pathFile, nameFileImag01)

}
