% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgWidthHeightChannel.R
\name{imgWidthHeightChannel}
\alias{imgWidthHeightChannel}
\title{gets the width, height and channels number values from image file}
\usage{
imgWidthHeightChannel(pathFile, nameFileImag01)
}
\arguments{
\item{pathFile}{: string (path)}

\item{nameFileImag01}{: string (file name)}
}
\value{
vector of numbers
}
\description{
gets the width, height and channels number values from image file
}
\examples{

pathFile <- "C:/Images"
nameFileImag01 <- "Image.bmp"
imgWidthHeightChannel(pathFile, nameFileImag01)

}
