% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadImagSelect.R
\name{loadImagSelect}
\alias{loadImagSelect}
\title{read image file}
\usage{
loadImagSelect(pathFile, nameFileImag)
}
\arguments{
\item{pathFile}{: string (path)}

\item{nameFileImag}{: string (file name)}
}
\value{
NULL or image file
}
\description{
read image file
}
\examples{

pathFile <- "C:/Images"
nameFileImag <- "Image.bmp"
loadImagSelect(pathFile, nameFileImag)

}
