% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTableInfoImg.R
\name{loadTableInfoImg}
\alias{loadTableInfoImg}
\title{read CSV file}
\usage{
loadTableInfoImg(pathFileCsv, nameFileCsv)
}
\arguments{
\item{pathFileCsv}{: string (path CSV file)}

\item{nameFileCsv}{: string (name CSV file)}
}
\value{
NULL or data.frame
}
\description{
read CSV file
}
\examples{

pathFileCsv <- "C:/Images"
nameFileCsv <- "Image.bmp"
loadTableInfoImg(pathFileCsv, nameFileCsv)

}
