% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadThisImag.R
\name{loadThisImag}
\alias{loadThisImag}
\title{try to read an image file}
\usage{
loadThisImag(newPathFile)
}
\arguments{
\item{newPathFile}{: string (path)}
}
\value{
NULL or object of class 'cimg'
}
\description{
try to read an image file
}
\examples{

newPathFile <- "C:/Images"
loadThisImag(newPathFile)

}
