% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTableInfoImg.R
\name{saveTableInfoImg}
\alias{saveTableInfoImg}
\title{save data.frame}
\usage{
saveTableInfoImg(pathFileCsv, nameFileCsv, dataCsv)
}
\arguments{
\item{pathFileCsv}{: string (path CSV file)}

\item{nameFileCsv}{: string (name CSV file)}

\item{dataCsv}{: data.frame}
}
\value{
: TRUE or FALSE
}
\description{
save data.frame
}
\examples{

vecNameFileImg <- c("Imagen", "Entropy", "Channels")
dataCsv <- data.frame(vecNameFileImg)

pathFileCsv <- "C:/CSV"
nameFileCsv <- "tableInfo.csv"
saveTableInfoImg(pathFileCsv, nameFileCsv, dataCsv)

}
