% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryGrayScale.R
\name{tryGrayScale}
\alias{tryGrayScale}
\title{try to convert from color model to grayscale model for raster images}
\usage{
tryGrayScale(image01, nameFileImag)
}
\arguments{
\item{image01}{: object of class 'cimg'}

\item{nameFileImag}{: string (file name)}
}
\value{
object grayscale image (spectrum == 1)
}
\description{
try to convert from color model to grayscale model for raster images
}
\examples{

image01 <- imager::cimg(array(1,c(10,10,5,3)))
nameFileImag <- "Image.bmp"
tryGrayScale(image01, nameFileImag)

}
