% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewGraphicInfo.R
\name{viewGraphicInfo}
\alias{viewGraphicInfo}
\title{shows graphics from image data frame}
\usage{
viewGraphicInfo(imgDataFrameLoc, isData, typeGraf)
}
\arguments{
\item{imgDataFrameLoc}{: data.frame}

\item{isData}{: numeric (data.frame column or field)}

\item{typeGraf}{: numeric (graphic type)}
}
\value{
0 or graphic
}
\description{
shows graphics from image data frame
}
\examples{

vecNameFileImg <- c("Imagen", "Entropy", "Std deviation", "Width", "Height", "Channels", "Min value", "Max value")
imgDataFrameLoc <- data.frame(vecNameFileImg)
isData <- TRUE
typeGraf <- TRUE
viewGraphicInfo(imgDataFrameLoc, isData, typeGraf)

}
