% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFileFormpathFileValid.R
\name{listFileFormpathFileValid}
\alias{listFileFormpathFileValid}
\title{creates an existing files names list of a directory}
\usage{
listFileFormpathFileValid(pathFile)
}
\arguments{
\item{pathFile}{: string (path)}
}
\value{
TRUE or FALSE , string vector or NULL
}
\description{
creates an existing files names list of a directory
}
